refactor(store): add Websocket plugin to Vuex store
